requirements.txt
streamlit==1.32.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
gspread==6.0.2
pandas==2.1.4
reportlab==4.0.7
pytz==2024.1
streamlit-lottie==0.0.5
Pillow==10.2.0
openpyxl==3.1.2
numpy==1.26.4
python-dateutil==2.8.2
requests==2.31.0
oauth2client==4.1.3
httplib2==0.22.0